+++ cypress/e2e/spec-wc-pyodide.cy.js
@@ -19,6 +19,12 @@ const runCode = (code) => {
 describe("Running the code with pyodide", () => {
   beforeEach(() => {
     cy.visit(origin);
+    cy.window().then((win) => {
+      Object.defineProperty(win, "crossOriginIsolated", {
+        value: true,
+        configurable: true,
+      });
+    });
   });
 
   it("runs a simple program", () => {