+++ src/components/Editor/Runners/PythonRunner/SkulptRunner/SkulptRunner.jsx
@@ -303,7 +303,7 @@ const SkulptRunner = ({ active, outputPanels = ["text", "visual"] }) => {
       const lineNumber = err.traceback[0].lineno;
       const fileName = err.traceback[0].filename.replace(/^\.\//, "");
 
-      if (errorType === "ImportError") {
+      if (errorType === "ImportError" && window.crossOriginIsolated) {
         const moduleName = errorDescription.replace(/No module named /, "");
         explanation = `You should check your code for typos. If you are using p5, py5, sense_hat or turtle, ${moduleName} might not work.`;
       }